Разбивает на модули файлы js
