+++ js/util.js
@@ -1,0 +1,26 @@
+/**
+ * Выберет по индексу случайное значение, в рамках заданного массива
+ * @template Item
+ * @param {Item[]} list
+ * @return {Item}
+ */
+const pickItemFromList = (list) => {
+  const index = Math.floor(Math.random() * list.length);
+
+  return list[index];
+};
+
+/**
+ * Выбирает случайное число в заданном диапазоне
+ * @param {number} min
+ * @param {number} max
+ * @return {number}
+ */
+const pickIntegerInRange = (min, max) => {
+  const value = min + Math.random() * (max - min);
+
+  return Math.round(value);
+};
+
+export {pickItemFromList};
+export {pickIntegerInRange};
